CXR-VLM/requirements.txt
numpy~=1.24.2
scikit-learn~=1.2.1
transformers~=4.27.4
# torchvision~=0.14.1
tqdm~=4.64.1
pandas~=1.5.3
pillow~=9.4.0
kornia~=0.6.10
pydicom==2.4.4
